Рефактор структуры проекта
