+++ src/parser/parse_TRS_and_apply_interpretations.jl
@@ -1,7 +1,13 @@
+module Parser
+
+include("types.jl")
+
 using Symbolics
 
+export parse_and_interpret, separatevars
+
 ############################ Функция для применения интерпретаций
-function apply_interpretation(term::Term, interpretations::Dict{String, Function}, var_map::Dict{String, String})
+function apply_interpretation(term, interpretations, var_map)::String
     if isempty(term.childs)
         # Если это переменная, возвращаем ее имя
         return term.name
@@ -31,7 +37,11 @@ function renamevars!(jsonterm1, jsonterm2, renamefunc)
     foreach(rename!, (jsonterm1, jsonterm2))
 end
 
-function separatevars(json_string)
+"""
+Разделяет переменные в каждом выражении, добавляя
+индекс. Возвращает json-строку 
+"""
+function separatevars(json_string)::String
     json_exprs = JSON.parse(json_string)
     for (index, expr) ∈ enumerate(json_exprs)
         renamevars!(expr["left"], expr["right"], x -> "$(x)$index")
@@ -121,4 +131,46 @@ function parse_and_interpret(json_string::String, interpretations::Dict{String, 
         # Сохраняем левую часть выражения
         push!(simplified_left_parts, string(difference_expanded))
     end
+end
+
+########################## Функция для парсинга интерпретаций
+function parse_interpretations(interpretations::Dict{String, Function}, var_map::Dict{String, String})
+    parsed_data = JSON.parse(json_interpret_string)
+
+    for func in parsed_data["functions"]
+        func_name = func["name"]
+        variables = func["variables"]
+        expression = func["expression"]
+
+        # Создаем функцию с необходимым количеством переменных
+        interpretations[func_name] = (vars...) -> begin
+            expr = expression
+            # Заменяем переменные в выражении на соответствующие переменные из TRS
+            for (i, var) in enumerate(variables)
+                # Используем регулярные выражения для замены целых слов
+                expr = replace(expr, Regex("\\b$(var)\\b") => vars[i])
+            end
+            return expr
+        end
+    end
+end
+
+########################################## Функция для парсинга термов
+# Функция для парсинга термов из JSON
+function parse_term(json::Dict)
+    childs = [parse_term(child) for child in json["childs"]]
+    return Term(json["value"], childs)
+end
+
+# Функция для отображения терма в человекочитаемом виде
+function term_to_string(term::Term)
+    if isempty(term.childs)
+        return term.name  # Если терм — переменная, возвращаем его имя
+    else
+        # Рекурсивно обрабатываем дочерние термы
+        child_strings = [term_to_string(child) for child in term.childs]
+        return "$(term.name)(" * join(child_strings, ", ") * ")"
+    end
+end
+
 end
